Rename fs_core.ts to fs_mem.ts
