+++ src/fs_mem.ts
