+++ src/index.ts
@@ -9,7 +9,7 @@ export {
   OpenDirectory,
   PreopenDirectory,
   ConsoleStdout,
-} from "./fs_core.js";
+} from "./fs_mem.js";
 export { SyncOPFSFile, OpenSyncOPFSFile } from "./fs_opfs.js";
 export { strace } from "./strace.js";
 export * as wasi from "./wasi_defs.js";
